move randomString() to a separate file, allowing build on macos
